requirements.txt
flask
Pillow
opencv-python
pyzxing